+++ SimplePerceptron/Perceptron.h
@@ -12,10 +12,14 @@ class Perceptron
 public:
 	Perceptron(int input_size);
 
+	//Train the perceptron with input data and a target label, this step will tweak the weights
+	//of the perceptron
 	void train(const std::vector<float>& inputs, int target);
 
-	float sign(float num);
+	//Returns -1 if num is less than 0, other wise return 1 
+	int sign(float num);
 
+	//Given a set of inputs, produce a prediction
 	int predict(const std::vector<float>& inputs);
 private:
 	int guess(const std::vector<float>& inputs);
